refactored popup into its own folder
