--- src/popup.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <link rel="stylesheet" type="text/css" href="styles/popup.css" />
-    <script src="./popup.js"></script>
-  </head>
-
-  <body>
-    <input type="file" id="input" />
-    <!-- <input type="button" id="button" value="Upload Keyfile" /> -->
-    <div>Main Functions</div>
-    <button onClick="browsingDataToLocalMirror()">Store Browsing Data To Local Mirror</button>
-    <button onClick="injectTestData()">Inject Test Data Into Local Mirror</button>
-    <button onClick="localMirrorToIPFS()">Push Local Mirror to IPFS</button>
-    <button onClick="setNewIPNSLink()">Set New IPNS Link</button>
-    <button onClick="fetchRemoteMirror()">Fetch Remote Mirror</button>
-    <button onClick="mergeMirrors()">Merge Mirrors</button>
-    <button onClick="loadMetadataIntoBrowser()">Update Browser Metadata</button>
-    <div>Misc</div>
-    <button onClick="">IPFS Status</button>
-  </body>
-</html>
