--- src/popup.js
@@ -1,29 +1,0 @@
-const browser = require('webextension-polyfill')
-
-const browsingDataToLocalMirror = () => {
-
-}
-
-const injectTestData = () => {
-
-}
-
-const localMirrorToIPFS = () => {
-
-}
-
-const setNewIPNSLink = () => {
-
-}
-
-const fetchRemoteMirror = () => {
-
-}
-
-const mergeMirrors = () => {
-
-}
-
-const loadMetadataIntoBrowser = () => {
-
-}
