+++ src/popup/popup.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" type="text/css" href="styles/popup.css" />
+  </head>
+
+  <body>
+    <!-- <input type="button" id="button" value="Upload Keyfile" /> -->
+    <div>Main Functions</div>
+    <button id="browsingDataToLocalMirror">Store Browsing Data To Local Mirror</button>
+    <button id="injectTestData">Inject Test Data Into Local Mirror</button>
+    <button id="localMirrorToIPFS">Push Local Mirror to IPFS</button>
+    <button id="setNewIPNSLink">Set New IPNS Link</button>
+    <button id="fetchRemoteMirror">Fetch Remote Mirror</button>
+    <button id="mergeMirrors">Merge Mirrors</button>
+    <button id="loadMetadataIntoBrowser">Update Browser Metadata</button>
+    <div>Misc</div>
+    <button onClick="">IPFS Status</button>
+    <script src="./popup.js"></script>
+  </body>
+</html>
